feat: 라벨 페이지 렌더링을 위해 클래스 생성 및 이슈 페이지 관련 코드 별도 파일로 이동
