+++ src/pages/label.js
@@ -1,0 +1,21 @@
+import { getLabelTpl } from '../tpl';
+
+export const LabelPage = class {
+
+  get template () {
+    return getLabelTpl();
+  }
+
+  constructor () {
+
+    this.initTemplate();
+  }
+
+  initData () {
+
+  }
+
+  initTemplate () {
+    createApp(this.template);
+  }
+}
